No permitir pasar el estado a Cargado de un pedido si la fecha que ingresa es menor a la de hoy

+++ Proyecto_EZ/Negocios/BusinessControllers/PedidoCtrl.cs
@@ -28,7 +28,7 @@ namespace Negocios.BusinessControllers
         #endregion
 
         #region COMPORTAMIENTO
-            
+
         public void GuardarPedido(int xiPedido, DateTime xdFechaEntrega, string xsEstado, int xiCliente, string xsUsuario, decimal xdTotal, decimal xdFacturado, List<Pedido> xoProductos, out string xsError)
         {
             xsError = "";
@@ -42,18 +42,28 @@ namespace Negocios.BusinessControllers
 
                     if (xoPedido != null)
                     {
-                        //Actualizamos el pedido
-                        xoPedido.ped_cliente = xiCliente;
-                        xoPedido.ped_monto = Math.Round(xdTotal);
-                        xoPedido.ped_factu = Math.Round(xdFacturado);
-                        xoPedido.ped_estado = xsEstado;
-                        xoPedido.ped_repartidor = xsUsuario;
+                        if (xoPedido.ped_estado.Trim().Equals("E") &&
+                            DateTime.Today > xdFechaEntrega &&
+                            xsEstado.Equals("C"))
+                        {
+                            xsError = "Si el pedido pasa a estado CARGADO, la fecha de entrega no puede ser menor a " + DateTime.Today.ToString("dd/MM/yyy");
+                        }
+                        else
+                        {
+                            //Actualizamos el pedido
+                            xoPedido.ped_cliente = xiCliente;
+                            xoPedido.ped_monto = Math.Round(xdTotal);
+                            xoPedido.ped_factu = Math.Round(xdFacturado);
+                            xoPedido.ped_estado = xsEstado;
+                            xoPedido.ped_repartidor = xsUsuario;
+                            xoPedido.ped_fecha = xdFechaEntrega;
 
-                        //Borramos todos los items del pedido
-                        var items = xoDB.pedido_detalle.Where(x => x.det_pedido == xiPedido).ToList();
-                        xoDB.pedido_detalle.RemoveRange(items);
+                            //Borramos todos los items del pedido
+                            var items = xoDB.pedido_detalle.Where(x => x.det_pedido == xiPedido).ToList();
+                            xoDB.pedido_detalle.RemoveRange(items);
 
-                        xoDB.SaveChanges();
+                            xoDB.SaveChanges();
+                        }
                     }
                     else
                     {
@@ -77,22 +87,25 @@ namespace Negocios.BusinessControllers
 
                     }
 
-                    //Para cada producto guardamos el detalle
-                    foreach (var item in xoProductos)
+                    if (xsError == "")
                     {
-                        var xoPedidoDetalle = new pedido_detalle()
+                        //Para cada producto guardamos el detalle
+                        foreach (var item in xoProductos)
                         {
-                            det_pedido = xoPedido != null ? xiPedido : xiIdPedido,
-                            det_producto = item.IdProducto,
-                            det_cantidad = item.Cantidad,
-                            det_precio = item.Precio,
-                            det_monto = item.Monto
-                        };
+                            var xoPedidoDetalle = new pedido_detalle()
+                            {
+                                det_pedido = xoPedido != null ? xiPedido : xiIdPedido,
+                                det_producto = item.IdProducto,
+                                det_cantidad = item.Cantidad,
+                                det_precio = item.Precio,
+                                det_monto = item.Monto
+                            };
 
-                        xoDB.pedido_detalle.Add(xoPedidoDetalle);
+                            xoDB.pedido_detalle.Add(xoPedidoDetalle);
+                        }
+
+                        xoDB.SaveChanges();
                     }
-
-                    xoDB.SaveChanges();
                 }
                 catch (Exception ex)
                 {
@@ -122,7 +135,7 @@ namespace Negocios.BusinessControllers
                         item.PedidoDetalle = ObtenerDetallePedido(item.IdPedido, out xsError);
                     }
 
-                    xoPedidos = xoResultado.Select(x => new [] {
+                    xoPedidos = xoResultado.Select(x => new[] {
                         x.IdPedido.ToString(),
                         x.Cliente + " - " + x.ClienteDireccion,
                         x.Fecha.ToString("dd-MM-yyyy"),
@@ -131,8 +144,8 @@ namespace Negocios.BusinessControllers
                         string.Format("{0:0.##}", x.Facturado),
                         x.Repartidor,
                         x.Estado,
-                        x.IdCliente.ToString(), 
-                        x.IdRepartidor,                       
+                        x.IdCliente.ToString(),
+                        x.IdRepartidor,
                         JsonConvert.SerializeObject(x.PedidoDetalle)}).ToList();
                 }
                 catch (Exception ex)
@@ -237,7 +250,7 @@ namespace Negocios.BusinessControllers
         public List<spGetFacturacionMensual> ObtenerFacturacionMensualRpt(DateTime xdFechaDesde, DateTime xdFechaHasta, string xsRepartidor)
         {
             var xoResultado = new List<spGetFacturacionMensual>();
-            if (xsRepartidor == "") xsRepartidor = null; 
+            if (xsRepartidor == "") xsRepartidor = null;
 
             using (BD_Entities xoDB = new BD_Entities())
             {
